Allow crane pull image to fail if older version is available (#10)
Sometimes pulling images fails due to service outages. When `pullOnRun` is used, an older image could still be available to use. If that is the case, there is no reason to refuse execution of tides. This change allows salvage to run tides even if it failed to pull the tides crane image. The tide itself will still be considered failed, since it goes against the user's intent to run old cranes when updates are explicitly required.